form-utils: Removed storybook.
It was not working, and there was only one basic story showcasing all of the components.

If we are to pay the dependency cost of storybook I think the reward needs to be bigger, so removing it for now.

Lets add it back in a way that works later, if we feel the need.